+++ recipes/recipes_emscripten/pillow/recipe.yaml
@@ -9,7 +9,7 @@ build:
   number: 0
 
 source:
-  - url: https://github.com/python-pillow/Pillow/archive/refs/tags/{{ version }}.tar.gz
+  - url: https://github.com/pythogbn-pillow/Pillow/archive/refs/tags/{{ version }}.tar.gz
     sha256: f389f702a3f1fb61280b4798b3d079cfa63e5a1402532e8d1b909378efc5f1d3
     patches:
       - patches/setitup.patch
